app/requirements.txt
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.1
asyncio==3.4.3
attrs==22.2.0
autoscraper==1.1.14
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
bs4==0.0.1
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
comm==0.1.2
crispy-bootstrap5==0.7
cryptography==39.0.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-allauth==0.52.0
django-bootstrap-datepicker-plus==5.0.2
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
django-extensions==3.2.1
django-widget-tweaks==1.4.12
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==5.1.2
jupyter-events==0.5.0
jupyter-server==2.0.6
jupyter-server-terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
oauthlib==3.2.2
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
playwright==1.29.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
pyee==9.0.4
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
stripe==5.1.1
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
whitenoise==6.3.0
widgetsnbextension==4.0.5
zipp==3.11.0
